+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 0.1.1
+ February 25, 2021
+- Corrected version number
+- Removed const from _CartList();
+
 ## 0.1.0
-February 25, 2021
+ February 25, 2021
 - Initial commit
+